# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Информация о текущем пользователе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить текущего пользователя",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Данные пользователя обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Public handle уже занят"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить данные текущего пользователя",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/profile": {
      "put": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профиль обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить профиль текущего пользователя",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{handle}": {
      "get": {
        "operationId": "UsersController_getPublicProfile",
        "parameters": [
          {
            "name": "handle",
            "required": true,
            "in": "path",
            "description": "Публичный handle пользователя",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Публичный профиль пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить публичный профиль по handle",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "PortfolioX API",
    "description": "Backend API для платформы PortfolioX - совместное резюме + портфолио",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Подтвержден ли email",
            "example": false
          },
          "name": {
            "type": "object",
            "description": "Имя пользователя",
            "nullable": true
          },
          "headline": {
            "type": "object",
            "description": "Заголовок профиля",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "description": "URL аватара",
            "nullable": true
          },
          "publicHandle": {
            "type": "object",
            "description": "Публичный handle",
            "nullable": true
          },
          "locale": {
            "type": "object",
            "description": "Локаль",
            "example": "en",
            "nullable": true
          },
          "profiles": {
            "description": "Профили пользователя",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileResponseDto"
            }
          },
          "_count": {
            "description": "Количество элементов",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserCountsDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "LinkResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID ссылки",
            "example": "clx1234567890"
          },
          "url": {
            "type": "string",
            "description": "URL ссылки",
            "example": "https://github.com/user"
          },
          "title": {
            "type": "object",
            "description": "Название ссылки",
            "example": "GitHub",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "url",
          "createdAt"
        ]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID профиля",
            "example": "clx1234567890"
          },
          "userId": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "bio": {
            "type": "object",
            "description": "Биография",
            "nullable": true
          },
          "visible": {
            "type": "boolean",
            "description": "Видимость профиля",
            "example": true
          },
          "links": {
            "description": "Ссылки профиля",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "visible",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserCountsDto": {
        "type": "object",
        "properties": {
          "resumes": {
            "type": "number",
            "description": "Количество резюме",
            "example": 3
          },
          "projects": {
            "type": "number",
            "description": "Количество проектов",
            "example": 5
          },
          "achievements": {
            "type": "number",
            "description": "Количество достижений",
            "example": 2
          }
        },
        "required": [
          "resumes",
          "projects",
          "achievements"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван Иванов"
          },
          "headline": {
            "type": "string",
            "description": "Заголовок профиля",
            "example": "Senior Full Stack Developer"
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL аватара пользователя",
            "example": "https://example.com/avatar.jpg"
          },
          "locale": {
            "type": "string",
            "description": "Локаль пользователя",
            "example": "ru-RU"
          },
          "publicHandle": {
            "type": "string",
            "description": "Публичный handle пользователя",
            "example": "ivan_ivanov"
          }
        }
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "description": "Биография пользователя",
            "example": "Опытный разработчик с 5+ годами опыта"
          },
          "visible": {
            "type": "boolean",
            "description": "Видимость профиля",
            "example": true
          }
        }
      },
      "PublicProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "name": {
            "type": "object",
            "description": "Имя пользователя",
            "nullable": true
          },
          "headline": {
            "type": "object",
            "description": "Заголовок профиля",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "description": "URL аватара",
            "nullable": true
          },
          "publicHandle": {
            "type": "object",
            "description": "Публичный handle",
            "nullable": true
          },
          "locale": {
            "type": "object",
            "description": "Локаль",
            "example": "en",
            "nullable": true
          },
          "profiles": {
            "description": "Профили пользователя",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileResponseDto"
            }
          },
          "resumes": {
            "type": "array",
            "description": "Опубликованные резюме"
          },
          "projects": {
            "type": "array",
            "description": "Публичные проекты"
          },
          "achievements": {
            "type": "array",
            "description": "Достижения"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}